requirements.txt
streamlit
sentence-transformers
openai
chromadb
pandas
python-dotenv

#.\venv\Scripts\activate
